- create model, migration, seeder customer search preference and all table relation

+++ database/migrations/2022_02_02_123936_add_plan_id_to_properties_table.php
@@ -27,9 +27,13 @@ class AddPlanIdToPropertiesTable extends Migration
      */
     public function down()
     {
-        Schema::table('properties', function (Blueprint $table) {
-            $table->dropForeign(['plan_id']);
-            $table->dropColumn('plan_id');
-        });
+        if (Schema::hasColumn('properties', 'plan_id')){
+            Schema::table('properties', function (Blueprint $table)
+            {
+                $table->dropForeign(['plan_id']);
+                $table->dropColumn('plan_id');
+            });
+        }
+
     }
 }
